Rework the core of the parsers
...such that we don't back-track by default.

I strongly suspect this was the source of the perf issues with the
unicode char db.

Most of the tests for the core are passing, except for the one for
`sepBy` -- the expression for which is returning a value that isn't even
the right type! I suspect this is this bug at work:

    https://github.com/unisonweb/unison/issues/693